Add DashboardPictureUploadModal
